PDB_1115

+++ src/main/java/com/ds/antddun/dto/JobListDTO.java
@@ -8,7 +8,7 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class JobListDTO {
-    private Long jno;
+    private int jno;
     private String job;
 
 }
